+++ content-scripts/src/modules/options/writerMode.js
@@ -5,7 +5,6 @@ import { createTypefullyLinkElement, createTypefullyLogo, getCurrentTextAndSendT
 import addStyles, { removeStyles, stylesExist } from "../utilities/addStyles";
 import addTooltip, { hideAllTooltips } from "../utilities/addTooltip";
 import addTypefullyBox from "../utilities/addTypefullyBox";
-import { createTypefullyUrl } from "../utilities/createTypefullyUrl";
 import { getStorage, setStorage } from "../utilities/storage";
 
 const escKeyListener = async (event) => {
@@ -130,21 +129,9 @@ export const addTypefullyPlugToWriterMode = async () => {
     typefullyLinkElement.appendChild(typefullyText);
 
     /* ----------------- Typefully box callout with explanation ---------------- */
-
-    const url = createTypefullyUrl({
-      utm_content: "writer-mode-callout",
-    });
-
     addTypefullyBox(
       main,
-      "writer-mode",
-      `<ul>
-  <li>💬 Share your drafts and get comments</li>
-  <li>🤖 Improve your tweets with AI</li>
-  <li>📈 Track your growth with insights and metrics</li>
-  <li>📆 Schedule for later</li>
-</ul>
-<p>Powered by <a href="${url}" target="_blank">Typefully</a>, the makers of the Minimal Twitter extension.</p>`,
+      "writer-mode-callout",
       {
         withArrow: true,
       }
@@ -158,7 +145,7 @@ export const removeTypefullyPlugFromWriterMode = () => {
   const typefullyLinkElement = document.getElementById("typefully-writermode-link");
   typefullyLinkElement && typefullyLinkElement.remove();
 
-  const typefullyBox = document.getElementById("typefully-writermode-box");
+  const typefullyBox = document.getElementById("typefully-callout-box");
   typefullyBox && typefullyBox.remove();
 };
