+++ content-scripts/src/modules/utilities/addTypefullyBox.js
@@ -1,16 +1,29 @@
 import svgAssets from "../svgAssets";
 import { getStorage, setStorage } from "./storage";
+import { createTypefullyUrl } from "./createTypefullyUrl";
 
-export default async function addTypefullyBox(rootElement, storageKey, innerHTML, options = {}) {
+export default async function addTypefullyBox(rootElement, utmContent, options = {}) {
   const { withArrow } = options ?? {};
 
-  const key = "tp-box-seen:" + storageKey;
+  const key = "tp-box-seen:typefully-callout";
 
   const seen = await getStorage(key);
 
+  const url = createTypefullyUrl({
+    utm_content: utmContent,
+  });
+
+  const innerHTML = `<ul>
+  <li>💬 Share your drafts and get comments</li>
+  <li>🤖 Improve your tweets with AI</li>
+  <li>📈 Track your growth with insights and metrics</li>
+  <li>📆 Schedule for later</li>
+</ul>
+<p>Powered by <a href="${url}" target="_blank">Typefully</a>, the makers of the Minimal Twitter extension.</p>`;
+
   if (seen !== "true") {
     const typefullyBox = document.createElement("div");
-    typefullyBox.id = "typefully-writermode-box";
+    typefullyBox.id = "typefully-callout-box";
     typefullyBox.className = "typefully-box";
 
     typefullyBox.innerHTML = innerHTML;
